+++ lib/app/app.dart
@@ -6,13 +6,13 @@ export 'services/storage.dart';
 export 'services/user_fetcher.dart';
 export 'theming_utils.dart';
 export 'utils.dart';
+export 'widgets/app_bar.dart';
 export 'widgets/buttons.dart';
+export 'widgets/card.dart';
 export 'widgets/empty_state.dart';
 export 'widgets/error_widgets.dart';
 export 'widgets/fade_in.dart';
-export 'widgets/fancy_app_bar.dart';
-export 'widgets/fancy_card.dart';
-export 'widgets/menu.dart' show NavigationItem;
+export 'widgets/navigation_item.dart';
 export 'widgets/page_route.dart' show TopLevelPageRoute;
 export 'widgets/placeholder_text.dart';
 export 'widgets/scaffold.dart';
