+++ lib/dashboard/widgets/dashboard_screen.dart
@@ -8,7 +8,7 @@ class DashboardScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return FancyScaffold(
-      appBar: FancyAppBar(title: Text('Dashboard')),
+      appBar: FancyAppBar(title: Text(context.s.dashboard)),
       sliver: SliverList(
         delegate: SliverChildListDelegate([
           CalendarDashboardCard(),
