+++ lib/l10n/intl_de.arb
@@ -28,29 +28,29 @@
   "@app_errorScreen_unknown_showStackTrace": {
     "type": "text"
   },
-  "app_navigation_assignments": "Hausaufgaben",
-  "@app_navigation_assignments": {
-    "type": "text"
+  "app_logOut_confirm": "Abmelden",
+  "@app_logOut_confirm": {
+    "type": "text",
+    "description": ""
   },
-  "app_navigation_courses": "Kurse",
-  "@app_navigation_courses": {
-    "type": "text"
+  "app_logOut_content": "Du brauchst deine Anmeldedaten, um dich wieder anzumelden.",
+  "@app_logOut_content": {
+    "type": "text",
+    "description": ""
   },
-  "app_navigation_dashboard": "Übersicht",
-  "@app_navigation_dashboard": {
-    "type": "text"
-  },
-  "app_navigation_files": "Dateien",
-  "@app_navigation_files": {
-    "type": "text"
-  },
-  "app_navigation_news": "Neuigkeiten",
-  "@app_navigation_news": {
-    "type": "text"
+  "app_logOut_title": "Abmelden?",
+  "@app_logOut_title": {
+    "type": "text",
+    "description": ""
   },
   "app_navigation_userDataEmpty": "—",
   "@app_navigation_userDataEmpty": {
     "type": "text"
+  },
+  "assignment": "Hausaufgaben",
+  "@assignment": {
+    "type": "text",
+    "description": ""
   },
   "assignment_assignmentsScreen_overdue": "Überfällig",
   "@assignment_assignmentsScreen_overdue": {
@@ -95,6 +95,11 @@
     "type": "text",
     "description": ""
   },
+  "course": "Kurse",
+  "@course": {
+    "type": "text",
+    "description": ""
+  },
   "course_coursesScreen_empty": "Es wurden keine Kurse gefunden.",
   "@course_coursesScreen_empty": {
     "type": "text"
@@ -102,6 +107,16 @@
   "course_detailsScreen_empty": "Dieser Kurs enthält noch keine Themen.",
   "@course_detailsScreen_empty": {
     "type": "text"
+  },
+  "dashboard": "Übersicht",
+  "@dashboard": {
+    "type": "text",
+    "description": ""
+  },
+  "file": "Dateien",
+  "@file": {
+    "type": "text",
+    "description": ""
   },
   "file_fileBrowser_downloading": "{fileName} wird heruntergeladen…",
   "@file_fileBrowser_downloading": {
@@ -138,6 +153,31 @@
   "file_files_my_description": "Hier findest du alle deine persönlichen Dateien.\nAuf diese Dateien hast nur du Zugriff, du kannst sie aber auch mit anderen Nutzern teilen.",
   "@file_files_my_description": {
     "type": "text"
+  },
+  "file_fileTile_details_createdAt": "Erstellt: {createdAt}",
+  "@file_fileTile_details_createdAt": {
+    "type": "text",
+    "description": ""
+  },
+  "file_fileTile_details_modifiedAt": "Zuletzt geändert: {modifiedAt}",
+  "@file_fileTile_details_modifiedAt": {
+    "type": "text",
+    "description": ""
+  },
+  "file_fileTile_details_offline": "Offline verfügbar machen",
+  "@file_fileTile_details_offline": {
+    "type": "text",
+    "description": ""
+  },
+  "file_fileTile_details_open": "Öffnen",
+  "@file_fileTile_details_open": {
+    "type": "text",
+    "description": ""
+  },
+  "general_cancel": "Abbrechen",
+  "@general_cancel": {
+    "type": "text",
+    "description": ""
   },
   "general_loading": "Lädt…",
   "@general_loading": {
@@ -206,6 +246,11 @@
   "login_loginScreen_placeholder": "Hier könnten noch weitere Informationen stehen.",
   "@login_loginScreen_placeholder": {
     "type": "text"
+  },
+  "news": "Neuigkeiten",
+  "@news": {
+    "type": "text",
+    "description": ""
   },
   "news_articlePreview_subtitle": "Veröffentlicht am {published} von {author}",
   "@news_articlePreview_subtitle": {
