+++ lib/l10n/intl_en.arb
@@ -28,29 +28,29 @@
   "@app_errorScreen_unknown_showStackTrace": {
     "type": "text"
   },
-  "app_navigation_assignments": "Assignments",
-  "@app_navigation_assignments": {
-    "type": "text"
+  "app_logOut_confirm": "Log out",
+  "@app_logOut_confirm": {
+    "type": "text",
+    "description": ""
   },
-  "app_navigation_courses": "Courses",
-  "@app_navigation_courses": {
-    "type": "text"
+  "app_logOut_content": "You'll need your login credentials to log in again.",
+  "@app_logOut_content": {
+    "type": "text",
+    "description": ""
   },
-  "app_navigation_dashboard": "Dashboard",
-  "@app_navigation_dashboard": {
-    "type": "text"
-  },
-  "app_navigation_files": "Files",
-  "@app_navigation_files": {
-    "type": "text"
-  },
-  "app_navigation_news": "News",
-  "@app_navigation_news": {
-    "type": "text"
+  "app_logOut_title": "Log out?",
+  "@app_logOut_title": {
+    "type": "text",
+    "description": ""
   },
   "app_navigation_userDataEmpty": "—",
   "@app_navigation_userDataEmpty": {
     "type": "text"
+  },
+  "assignment": "Assignments",
+  "@assignment": {
+    "type": "text",
+    "description": ""
   },
   "assignment_assignmentsScreen_overdue": "Overdue",
   "@assignment_assignmentsScreen_overdue": {
@@ -95,6 +95,11 @@
     "type": "text",
     "description": ""
   },
+  "course": "Courses",
+  "@course": {
+    "type": "text",
+    "description": ""
+  },
   "course_coursesScreen_empty": "Seems like you're currently not enrolled in any courses.",
   "@course_coursesScreen_empty": {
     "type": "text"
@@ -102,6 +107,16 @@
   "course_detailsScreen_empty": "This course doesn't contain any topics.",
   "@course_detailsScreen_empty": {
     "type": "text"
+  },
+  "dashboard": "Dashboard",
+  "@dashboard": {
+    "type": "text",
+    "description": ""
+  },
+  "file": "Files",
+  "@file": {
+    "type": "text",
+    "description": ""
   },
   "file_fileBrowser_downloading": "Downloading {fileName}…",
   "@file_fileBrowser_downloading": {
@@ -138,6 +153,31 @@
   "file_files_my_description": "These are your personal files.\nBy default, only you can access them, but they may be shared with others.",
   "@file_files_my_description": {
     "type": "text"
+  },
+  "file_fileTile_details_createdAt": "Created: {createdAt}",
+  "@file_fileTile_details_createdAt": {
+    "type": "text",
+    "description": ""
+  },
+  "file_fileTile_details_modifiedAt": "Last modified: {modifiedAt}",
+  "@file_fileTile_details_modifiedAt": {
+    "type": "text",
+    "description": ""
+  },
+  "file_fileTile_details_offline": "Make available offline",
+  "@file_fileTile_details_offline": {
+    "type": "text",
+    "description": ""
+  },
+  "file_fileTile_details_open": "Open",
+  "@file_fileTile_details_open": {
+    "type": "text",
+    "description": ""
+  },
+  "general_cancel": "Cancel",
+  "@general_cancel": {
+    "type": "text",
+    "description": ""
   },
   "general_loading": "Loading…",
   "@general_loading": {
@@ -206,6 +246,11 @@
   "login_loginScreen_placeholder": "There could go some other information down here.",
   "@login_loginScreen_placeholder": {
     "type": "text"
+  },
+  "news": "News",
+  "@news": {
+    "type": "text",
+    "description": ""
   },
   "news_articlePreview_subtitle": "published on {published} by {author}",
   "@news_articlePreview_subtitle": {
